+++ database/migrations/2018_06_07_105701_create_question_forms_table.php
@@ -13,43 +13,45 @@ class CreateQuestionFormsTable extends Migration
      */
     public function up()
     {
-        Schema::create('tbl_question_forms', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('qfPId')->length(10)->unsigned();
-            $table->integer('qfFrId')->length(10)->unsigned();
-            $table->integer('qfQId')->length(10)->unsigned();
-            $table->integer('qfQsId')->length(10)->unsigned()->nullable();
-            $table->string('qfFaErrorMsg')->nullable();
-            $table->string('qfEnErrorMsg')->nullable();
-            $table->string('qfArErrorMsg')->nullable();
-            $table->unsignedTinyInteger('qfRequire')->default(1);
-            $table->unsignedTinyInteger('qfStep')->default(1);
-            $table->unsignedTinyInteger('qfGrid')->default(12);
-            $table->unsignedTinyInteger('qfUnique')->default(0);
-            $table->unsignedTinyInteger('qfOrder')->default(1);
-            $table->unsignedTinyInteger('qfState')->default(1);
-            $table->timestamps();
+        if (!Schema::hasTable('tbl_question_forms')) {
+            Schema::create('tbl_question_forms', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->bigInteger('qfPId')->length(20)->unsigned();
+                $table->bigInteger('qfFrId')->length(20)->unsigned();
+                $table->bigInteger('qfQId')->length(20)->unsigned();
+                $table->bigInteger('qfQsId')->length(20)->unsigned()->nullable();
+                $table->string('qfFaErrorMsg')->nullable();
+                $table->string('qfEnErrorMsg')->nullable();
+                $table->string('qfArErrorMsg')->nullable();
+                $table->unsignedTinyInteger('qfRequire')->default(1);
+                $table->unsignedTinyInteger('qfStep')->default(1);
+                $table->unsignedTinyInteger('qfGrid')->default(12);
+                $table->unsignedTinyInteger('qfUnique')->default(0);
+                $table->unsignedTinyInteger('qfOrder')->default(1);
+                $table->unsignedTinyInteger('qfState')->default(1);
+                $table->timestamps();
 
-            $table->foreign('qfPId')
-                ->references('id')->on('tbl_patterns')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
+                $table->foreign('qfPId')
+                    ->references('id')->on('tbl_patterns')
+                    ->onDelete('restrict')
+                    ->onUpdate('cascade');
 
-            $table->foreign('qfFrId')
-                ->references('id')->on('tbl_forms')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
+                $table->foreign('qfFrId')
+                    ->references('id')->on('tbl_forms')
+                    ->onDelete('restrict')
+                    ->onUpdate('cascade');
 
-            $table->foreign('qfQId')
-                ->references('id')->on('tbl_questions')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
+                $table->foreign('qfQId')
+                    ->references('id')->on('tbl_questions')
+                    ->onDelete('restrict')
+                    ->onUpdate('cascade');
 
-            $table->foreign('qfQsId')
-                ->references('id')->on('tbl_question_steps')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
-        });
+                $table->foreign('qfQsId')
+                    ->references('id')->on('tbl_question_steps')
+                    ->onDelete('restrict')
+                    ->onUpdate('cascade');
+            });
+        }
     }
 
     /**
